+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Change Log
 
-All notable changes to the "rails-buddy" extension will be documented in this file.
+All notable changes to the "erbalist" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 ### Added
-- New configuration option `rubyMarkup.highlightMode` to control Ruby highlighting behavior
+- New configuration option `erbalist.highlightMode` to control Ruby highlighting behavior
   - `always`: Always highlight Ruby code
   - `whenInBlock`: Only highlight Ruby code within Ruby blocks